+++ CHANGELOG.md
@@ -1,2 +1,6 @@
-## 0.1.0 (release date: 2014-02-23)
+## 0.1.1 (release date: 2015-03-02)
+  - new SSL certs directory
+  - bootstrap capabilities
+
+## 0.1.0 (release date: 2015-02-23)
   - Initial release